refactor(httpd): move grug-httpd scaffolding into indexer-httpd
Move the remaining grug-httpd scaffolding into indexer-httpd:

- `middlewares/shutdown.rs`, `subscription_limiter.rs`, `request_ip.rs`,
  `routes/graphql.rs`, `routes/index.rs` → relocated under `indexer/httpd/src/`.
- `metrics.rs` (with `GaugeGuard`) → moved; `init_graphql_metrics` renamed
  to `init_grug_query_metrics` to avoid collision with indexer-httpd's
  existing `graphql::extensions::metrics::init_graphql_metrics` (different
  histograms, same name).
- `traits.rs` (`QueryApp`) → merged into indexer-httpd's `traits.rs`,
  which already housed `ConsensusClient`.
- `error.rs` → grug-httpd's `Error` enum had `GraphQL` and
  `ActixWebMetricsBuilder` variants plus a `From<async_graphql::Error>`
  impl that were all dead code (no callers anywhere in the workspace);
  they are dropped, not carried over. indexer-httpd's existing `Error`
  stays unchanged.
- `context.rs` — `MinimalContext` is now a local struct in indexer-httpd
  (previously a `pub use` re-export of `grug_httpd::context::Context`).

Add a `minimal_up` handler so chain-only mode can serve `/up` against a
`MinimalContext` (returns `indexed_block_height: None`); the existing
`up` handler stays as the full-mode variant.

`futures-util` becomes a non-optional dep of indexer-httpd because the
moved `middlewares/shutdown.rs` and `subscription_limiter.rs` use
`futures_util::*` unconditionally.

Sweep imports: every `grug_httpd::*` reference across the workspace flips
to `indexer_httpd::*` (or `crate::*` for indexer-httpd-internal files).
`grug-httpd`'s Cargo.toml is trimmed to a no-content placeholder and all
callers drop the `grug-httpd` dependency.

`grug-httpd` is now an empty crate; commit 7 deletes it. SDL byte-identical.

Part of phase 3 of the indexer refactor.
