+++ dango/cli/Cargo.toml
@@ -23,7 +23,6 @@ dialoguer                   = { workspace = true }
 futures                     = { workspace = true }
 grug-app                    = { workspace = true, features = ["abci", "metrics", "tracing"] }
 grug-db-disk                = { workspace = true, features = ["metrics", "tracing"] }
-grug-httpd                  = { workspace = true, features = ["metrics", "tracing"] }
 grug-types                  = { workspace = true }
 grug-vm-rust                = { workspace = true }
 home                        = { workspace = true }
