+++ dango/mocks/httpd/Cargo.toml
@@ -19,7 +19,6 @@ dango-proposal-preparer  = { workspace = true }
 dango-testing            = { workspace = true }
 grug-app                 = { workspace = true }
 grug-db-memory           = { workspace = true }
-grug-httpd               = { workspace = true }
 grug-testing             = { workspace = true }
 grug-types               = { workspace = true }
 grug-vm-rust             = { workspace = true }
