+++ dango/testing/Cargo.toml
@@ -25,7 +25,6 @@ grug                     = { workspace = true }
 grug-app                 = { workspace = true }
 grug-db-disk             = { workspace = true }
 grug-db-memory           = { workspace = true }
-grug-httpd               = { workspace = true }
 grug-types               = { workspace = true, features = ["async-graphql"] }
 grug-vm-rust             = { workspace = true }
 hex-literal              = { workspace = true }
