+++ grug/httpd/Cargo.toml
@@ -9,37 +9,7 @@ repository    = { workspace = true }
 rust-version  = { workspace = true }
 version       = { workspace = true }
 
-[features]
-metrics = ["dep:actix-web-metrics", "dep:metrics"]
-testing = ["dep:grug-testing"]
-tracing = ["dep:tracing"]
+# All scaffolding moved into `indexer-httpd` in phase 3 of
+# testdata/refactor-indexer.md. This crate is deleted in commit 7.
 
 [dependencies]
-actix                   = { workspace = true }
-actix-cors              = { workspace = true }
-actix-web               = { workspace = true }
-actix-web-metrics       = { workspace = true, optional = true }
-anyhow                  = { workspace = true }
-async-graphql           = { workspace = true }
-async-graphql-actix-web = { workspace = true }
-async-trait             = { workspace = true }
-chrono                  = { workspace = true }
-futures-util            = { workspace = true }
-grug-app                = { workspace = true, features = ["tracing"] }
-grug-testing            = { workspace = true, optional = true }
-grug-types              = { workspace = true, features = ["async-graphql", "chrono", "tendermint"] }
-metrics                 = { workspace = true, optional = true }
-sentry-actix            = { workspace = true }
-serde                   = { workspace = true }
-serde_json              = { workspace = true }
-tempfile                = { workspace = true }
-tendermint              = { workspace = true }
-tendermint-rpc          = { workspace = true, features = ["http-client"] }
-thiserror               = { workspace = true }
-tokio                   = { workspace = true }
-tracing                 = { workspace = true, optional = true }
-uuid                    = { workspace = true }
-
-[dev-dependencies]
-clap               = { workspace = true }
-tracing-subscriber = { workspace = true }
