--- grug/httpd/src/error.rs
@@ -1,19 +1,0 @@
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum Error {
-    #[error("GraphQL error: {0}")]
-    GraphQL(String),
-
-    #[error("IO error: {0}")]
-    Io(#[from] std::io::Error),
-
-    #[error("Actix web metrics builder error")]
-    ActixWebMetricsBuilder,
-}
-
-impl From<async_graphql::Error> for Error {
-    fn from(e: async_graphql::Error) -> Self {
-        Error::GraphQL(format!("{e:?}"))
-    }
-}
