--- grug/httpd/src/middlewares/mod.rs
@@ -1,1 +1,0 @@
-pub mod shutdown;
