--- grug/httpd/src/routes.rs
@@ -1,2 +1,0 @@
-pub mod graphql;
-pub mod index;
