--- grug/httpd/src/routes/index.rs
@@ -1,52 +1,0 @@
-use {
-    crate::{context::Context, request_ip::RequesterIp},
-    actix_web::{
-        Error, HttpRequest, HttpResponse, Responder, error::ErrorInternalServerError, get, web,
-    },
-    async_graphql::futures_util::TryFutureExt,
-    chrono::{Duration, Utc},
-    grug_types::{BlockInfo, GIT_COMMIT},
-    std::env::var,
-};
-
-#[get("/")]
-pub async fn index() -> impl Responder {
-    "OK"
-}
-
-#[get("/requester-ip")]
-pub async fn requester_ip(req: HttpRequest) -> impl Responder {
-    HttpResponse::Ok().json(RequesterIp::from_request(&req))
-}
-
-#[derive(serde::Serialize)]
-pub struct UpResponse<'a> {
-    pub block: BlockInfo,
-    pub is_running: bool,
-    pub git_commit: &'a str,
-    pub indexed_block_height: Option<u64>,
-    pub chain_id: &'a str,
-    pub hostname: &'a str,
-}
-
-#[get("/up")]
-pub async fn up(app_ctx: web::Data<Context>) -> Result<impl Responder, Error> {
-    // This ensures that grug is working
-    let block = app_ctx
-        .grug_app
-        .last_finalized_block()
-        .map_err(ErrorInternalServerError)
-        .await?;
-
-    let is_running =
-        block.timestamp.to_naive_date_time() >= (Utc::now().naive_utc() - Duration::seconds(30));
-
-    Ok(HttpResponse::Ok().json(UpResponse {
-        block,
-        is_running,
-        git_commit: GIT_COMMIT,
-        indexed_block_height: None,
-        chain_id: var("CHAIN_ID").unwrap_or_default().as_str(),
-        hostname: var("HOSTNAME").unwrap_or_default().as_str(),
-    }))
-}
