--- grug/httpd/src/traits.rs
@@ -1,122 +1,0 @@
-#[cfg(feature = "testing")]
-use grug_testing::TestSuite;
-use {
-    async_trait::async_trait,
-    grug_app::{
-        App, AppError, AppResult, CHAIN_ID, Db, Indexer, LAST_FINALIZED_BLOCK, ProposalPreparer, Vm,
-    },
-    grug_types::{BlockInfo, Query, QueryResponse, TxOutcome, UnsignedTx},
-};
-
-#[async_trait]
-pub trait QueryApp {
-    /// Query the app, return the query response and the block height this query
-    /// was performed at.
-    async fn query_app(
-        &self,
-        raw_req: Query,
-        height: Option<u64>,
-    ) -> AppResult<(QueryResponse, u64)>;
-
-    /// Query the app's underlying key-value store, return `(value, proof, height)`.
-    async fn query_store(
-        &self,
-        key: &[u8],
-        height: Option<u64>,
-        prove: bool,
-    ) -> AppResult<(Option<Vec<u8>>, Option<Vec<u8>>, u64)>;
-
-    /// Simulate a transaction.
-    async fn simulate(&self, unsigned_tx: UnsignedTx) -> AppResult<TxOutcome>;
-
-    /// Query the chain ID.
-    async fn chain_id(&self) -> AppResult<String>;
-
-    /// Query the last finalized block.
-    async fn last_finalized_block(&self) -> AppResult<BlockInfo>;
-}
-
-#[async_trait]
-impl<DB, VM, PP, ID> QueryApp for App<DB, VM, PP, ID>
-where
-    DB: Db + Send + Sync + 'static,
-    VM: Vm + Clone + Send + Sync + 'static,
-    PP: ProposalPreparer + Send + Sync + 'static,
-    ID: Indexer + Send + Sync + 'static,
-    AppError: From<DB::Error> + From<VM::Error> + From<PP::Error>,
-{
-    async fn query_app(
-        &self,
-        raw_req: Query,
-        height: Option<u64>,
-    ) -> AppResult<(QueryResponse, u64)> {
-        Ok(self.do_query_app_with_height(raw_req, height, false)?)
-    }
-
-    async fn query_store(
-        &self,
-        key: &[u8],
-        height: Option<u64>,
-        prove: bool,
-    ) -> AppResult<(Option<Vec<u8>>, Option<Vec<u8>>, u64)> {
-        self.do_query_store_with_height(key, height, prove)
-    }
-
-    async fn simulate(&self, unsigned_tx: UnsignedTx) -> AppResult<TxOutcome> {
-        Ok(self.do_simulate(unsigned_tx, 0, false)?)
-    }
-
-    async fn chain_id(&self) -> AppResult<String> {
-        let storage = self.db.state_storage_with_comment(None, "httpd")?;
-        let chain_id = CHAIN_ID.load(&storage)?;
-
-        Ok(chain_id)
-    }
-
-    async fn last_finalized_block(&self) -> AppResult<BlockInfo> {
-        let storage = self.db.state_storage_with_comment(None, "httpd")?;
-        let last_finalized_block = LAST_FINALIZED_BLOCK.load(&storage)?;
-
-        Ok(last_finalized_block)
-    }
-}
-
-#[cfg(feature = "testing")]
-#[async_trait]
-impl<DB, VM, PP, ID> QueryApp for TestSuite<DB, VM, PP, ID>
-where
-    DB: Db + Send + Sync + 'static,
-    VM: Vm + Clone + Send + Sync + 'static,
-    PP: ProposalPreparer + Send + Sync + 'static,
-    ID: Indexer + Send + Sync + 'static,
-    App<DB, VM, PP, ID>: QueryApp,
-{
-    async fn query_app(
-        &self,
-        raw_req: Query,
-        height: Option<u64>,
-    ) -> AppResult<(QueryResponse, u64)> {
-        self.app.query_app(raw_req, height).await
-    }
-
-    async fn query_store(
-        &self,
-        key: &[u8],
-        height: Option<u64>,
-        prove: bool,
-    ) -> AppResult<(Option<Vec<u8>>, Option<Vec<u8>>, u64)> {
-        self.app.query_store(key, height, prove).await
-    }
-
-    async fn simulate(&self, unsigned_tx: UnsignedTx) -> AppResult<TxOutcome> {
-        self.app.simulate(unsigned_tx).await
-    }
-
-    async fn chain_id(&self) -> AppResult<String> {
-        self.app.chain_id().await
-    }
-
-    async fn last_finalized_block(&self) -> AppResult<BlockInfo> {
-        self.app.last_finalized_block().await
-    }
-}
