+++ indexer/httpd/src/graphql/subscription/block.rs
@@ -1,9 +1,9 @@
 #[cfg(feature = "metrics")]
-use grug_httpd::metrics::GaugeGuard;
+use crate::metrics::GaugeGuard;
 use {
+    crate::subscription_limiter::{acquire_subscription, guard_subscription_stream},
     async_graphql::{futures_util::stream::Stream, *},
     futures_util::stream::{StreamExt, once},
-    grug_httpd::subscription_limiter::{acquire_subscription, guard_subscription_stream},
     indexer_sql::entity,
     sea_orm::{ColumnTrait, EntityTrait, QueryFilter, QueryOrder},
     std::sync::{
