+++ indexer/httpd/src/graphql/subscription/grug.rs
@@ -1,14 +1,16 @@
+#[cfg(feature = "metrics")]
+use {crate::metrics::GaugeGuard, std::sync::Arc};
 use {
-    crate::graphql::{
-        query::grug::GrugQuery,
-        types::{query_response::QueryResponseWithBlockHeight, status::Status, store::Store},
+    crate::{
+        graphql::{
+            query::grug::GrugQuery,
+            types::{query_response::QueryResponseWithBlockHeight, status::Status, store::Store},
+        },
+        subscription_limiter::{acquire_subscription, guard_subscription_stream},
     },
     async_graphql::{futures_util::stream::Stream, *},
     futures_util::stream::{StreamExt, once},
-    grug_httpd::subscription_limiter::{acquire_subscription, guard_subscription_stream},
 };
-#[cfg(feature = "metrics")]
-use {grug_httpd::metrics::GaugeGuard, std::sync::Arc};
 
 #[derive(Default)]
 pub struct GrugSubscription;
