+++ indexer/httpd/src/graphql/subscription/transfer.rs
@@ -1,16 +1,18 @@
+#[cfg(feature = "metrics")]
+use {crate::metrics::GaugeGuard, std::sync::Arc};
 use {
-    crate::graphql::subscription::MAX_PAST_BLOCKS,
+    crate::{
+        graphql::subscription::MAX_PAST_BLOCKS,
+        subscription_limiter::{acquire_subscription, guard_subscription_stream},
+    },
     async_graphql::{futures_util::stream::Stream, *},
     dango_types::account_factory::UserIndex,
     futures_util::stream::{StreamExt, once},
-    grug_httpd::subscription_limiter::{acquire_subscription, guard_subscription_stream},
     indexer_sql::{entity, entity::blocks::latest_block_height},
     itertools::Itertools,
     sea_orm::{ColumnTrait, EntityTrait, QueryFilter, QueryOrder},
     std::ops::RangeInclusive,
 };
-#[cfg(feature = "metrics")]
-use {grug_httpd::metrics::GaugeGuard, std::sync::Arc};
 
 #[derive(Default)]
 pub struct TransferSubscription;
