+++ indexer/httpd/src/lib.rs
@@ -1,9 +1,13 @@
 pub mod context;
 pub mod error;
 pub mod graphql;
+#[cfg(feature = "metrics")]
+pub mod metrics;
 pub mod middlewares;
+pub mod request_ip;
 pub mod routes;
 pub mod server;
+pub mod subscription_limiter;
 mod tendermint_rpc_client;
 pub mod traits;
