+++ indexer/httpd/src/middlewares.rs
@@ -1,2 +1,3 @@
 #[cfg(feature = "metrics")]
 pub mod metrics;
+pub mod shutdown;
