+++ indexer/httpd/src/middlewares/shutdown.rs
