+++ indexer/httpd/src/routes.rs
@@ -1,2 +1,3 @@
 pub mod blocks;
+pub mod graphql;
 pub mod index;
