+++ indexer/testing/Cargo.toml
@@ -23,7 +23,6 @@ futures-util             = { workspace = true }
 graphql_client           = { workspace = true }
 grug-app                 = { workspace = true }
 grug-db-memory           = { workspace = true }
-grug-httpd               = { workspace = true }
 grug-testing             = { workspace = true }
 grug-types               = { workspace = true }
 grug-vm-rust             = { workspace = true }
